requirements.txt
# requirements.txt
fastapi
uvicorn
celery
redis
pydantic
requests
python-json-logger
tenacity
python-dotenv